requirements.txt
#apache-airflow>=1.10.0
ansible>=2.6.4
flask>=1.0.2
flask_cors>=3.0.6
flask_oslolog>=0.1
flask_restplus>=0.11.0
paramiko>=2.4.1
docker>=3.5.0
six>=1.10.0
sqlalchemy>=1.2.11
setuptools
pytest>=3.8.0
python-terraform>=0.10.0
oslo_db>=4.40.0
pywinrm>=0.3.0
jsonpickle>=0.9.6
scp>=0.11.0
mysqlclient>=1.3.13
pycrypto
pandas
xlrd
pytest-mock
pytest-flask
pytest-html
requests
pytest-easy-api
pycryptodomex
python-jenkins
oslo.messaging>=5.17.1
oslo.service>=1.16.0
oslo.utils>=3.22.0